# servers/assessment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the assessment server of PROMPT.",
        "title": "PROMPT Assessment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/assessment/api",
    "paths": {
...
        "/course_phase/{coursePhaseID}/evaluation": {
            "get": {
                "description": "List evaluations for a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "List evaluations by course phase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/evaluationDTO.Evaluation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create or update an evaluation for the current student.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "Create or update evaluation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Evaluation payload",
                        "name": "evaluation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/evaluationDTO.CreateOrUpdateEvaluationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/course_phase/{coursePhaseID}/evaluation/my-evaluations": {
            "get": {
                "description": "List evaluations authored by the current student.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "List my evaluations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/evaluationDTO.Evaluation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/evaluation/tutor/{courseParticipationID}": {
            "get": {
                "description": "List evaluations for a tutor in a course phase.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "evaluations"
                ],
                "summary": "List evaluations for tutor in phase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tutor course participation ID",
                        "name": "courseParticipationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/evaluationDTO.Evaluation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/course_phase/{coursePhaseID}/evaluation/{evaluationID}": {
            "delete": {
                "description": "Delete an evaluation by ID.",
                "tags": [
                    "evaluations"
                ],
                "summary": "Delete evaluation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course phase ID",
                        "name": "coursePhaseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Evaluation ID",
                        "name": "evaluationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "assessmentType.AssessmentType": {
            "type": "string",
            "enum": [
                "self",
                "peer",
                "tutor",
                "assessment"
            ],
            "x-enum-varnames": [
                "Self",
                "Peer",
                "Tutor",
                "Assessment"
            ]
        },
...
        "evaluationDTO.CreateOrUpdateEvaluationRequest": {
            "type": "object",
            "required": [
                "authorCourseParticipationID",
                "competencyID",
                "courseParticipationID",
                "scoreLevel",
                "type"
            ],
            "properties": {
                "authorCourseParticipationID": {
                    "type": "string"
                },
                "competencyID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "scoreLevel": {
                    "enum": [
                        "veryBad",
                        "bad",
                        "ok",
                        "good",
                        "veryGood"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/scoreLevelDTO.ScoreLevel"
                        }
                    ]
                },
                "type": {
                    "enum": [
                        "self",
                        "peer",
                        "tutor",
                        "assessment"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/assessmentType.AssessmentType"
                        }
                    ]
                }
            }
        },
        "evaluationDTO.Evaluation": {
            "type": "object",
            "properties": {
                "authorCourseParticipationID": {
                    "type": "string"
                },
                "competencyID": {
                    "type": "string"
                },
                "courseParticipationID": {
                    "type": "string"
                },
                "coursePhaseID": {
                    "type": "string"
                },
                "evaluatedAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "id": {
                    "type": "string"
                },
                "scoreLevel": {
                    "$ref": "#/definitions/scoreLevelDTO.ScoreLevel"
                },
                "type": {
                    "$ref": "#/definitions/assessmentType.AssessmentType"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "scoreLevelDTO.ScoreLevel": {
            "type": "string",
            "enum": [
                "veryBad",
                "bad",
                "ok",
                "good",
                "veryGood"
            ],
            "x-enum-varnames": [
                "ScoreLevelVeryBad",
                "ScoreLevelBad",
                "ScoreLevelOk",
                "ScoreLevelGood",
                "ScoreLevelVeryGood"
            ]
        },
...
    },
...
}